Added framework to do something with the new field that the master is supplying.
Changed the general workings of the script since we need to do a big, ugly
regexp at the beginning in order to s/,/./g in the email field anyway.

There's a lot of logic that's been added to allow us to turn this into a
'log processor' that can generate information useful to others in DCTI, but
I can't think of a good way to impliment it right now, so it's commented out
at the moment. For now, Daa is keeping an eye on bad blocks and what-not by
hand, but we should come up with a better means of handling that.