We end up updating every team that has work from new members, whether it's a new team or not, so update new and old teams the same (meaning set WORK_TOTAL in both instances).
Since WORK_TODAY is already contained in WORK_TOTAL, don't include it in the
amount of work used to adjust WORK_TOTAL for teams to account for the previous
work of new members.

Actually, some of this crap should go away now that newjoin.sql exists. That
script is really what should handle any prior work... if you join a team on a
day, that should take effect, even if you don't submit any work for that day.

All of team processing really needs a code review at this point.